fixes in parseConst; move getFunctionIndex into Runtime; test for global function pointers

+++ src/utility.js
@@ -224,6 +224,6 @@ function setIntersect(x, y) {
 }
 
 function isNumber(x) {
-  return x == parseInt(x);
+  return x == parseFloat(x);
 }
